@@ -1,7 +1,7 @@
 [package]
 name = "adv-code-2025"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 anyhow = "1.0.100"
