@@ -1,9 +1,12 @@
-use anyhow::*;
-use std::fs::File;
-use std::io::{BufRead, BufReader};
+#![allow(unexpected_cfgs)]
+#![warn(clippy::all, clippy::pedantic)]
+
+use adv_code_2025::start_day;
+use anyhow::{Ok, Result};
 use code_timing_macros::time_snippet;
 use const_format::concatcp;
-use adv_code_2025::*;
+use std::fs::File;
+use std::io::{BufRead, BufReader};
 
 const DAY: &str = "NN"; // TODO: Fill the day
 const INPUT_FILE: &str = concatcp!("input/", DAY, ".txt");
@@ -18,9 +21,10 @@ fn main() -> Result<()> {
     //region Part 1
     println!("=== Part 1 ===");
 
+    #[allow(clippy::items_after_statements)]
     fn part1<R: BufRead>(reader: R) -> Result<usize> {
         // TODO: Solve Part 1 of the puzzle
-        let answer = reader.lines().flatten().count();
+        let answer = reader.lines().map_while(Result::ok).count();
         Ok(answer)
     }
 
@@ -29,12 +33,13 @@ fn main() -> Result<()> {
 
     let input_file = BufReader::new(File::open(INPUT_FILE)?);
     let result = time_snippet!(part1(input_file)?);
-    println!("Result = {}", result);
+    println!("Result = {result}");
     //endregion
 
     //region Part 2
     // println!("\n=== Part 2 ===");
     //
+    // #[allow(clippy::items_after_statements)]
     // fn part2<R: BufRead>(reader: R) -> Result<usize> {
     //     Ok(0)
     // }
@@ -43,7 +48,7 @@ fn main() -> Result<()> {
     //
     // let input_file = BufReader::new(File::open(INPUT_FILE)?);
     // let result = time_snippet!(part2(input_file)?);
-    // println!("Result = {}", result);
+    // println!("Result = {result}");
     //endregion
 
     Ok(())
